+++ RELEASES.md
@@ -529,3 +529,13 @@ Based on:
 - [typescript v0.1.0-alpha.2] .
 ### Releases
 - [NPM v0.1.0-alpha.2] https://www.npmjs.com/package/cribl-control-plane/v/0.1.0-alpha.2 - .
+
+## 2025-10-03 07:15:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.634.1 (2.721.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.0-alpha.3] .
+### Releases
+- [NPM v0.1.0-alpha.3] https://www.npmjs.com/package/cribl-control-plane/v/0.1.0-alpha.3 - .